requirements.txt
requests==2.28.1
gradio==3.12.1
python-dotenv==1.0.0
